# contracts/dist/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Golem pilot contracts",
    "version": "5.26.0"
  },
  "components": {
    "schemas": {
      "def-0": {
        "anyOf": [
          {
            "type": "string"
          },
          {
            "type": "number"
          },
          {
            "type": "boolean"
          },
          {
            "type": "string",
            "nullable": true,
            "enum": [
              null
            ]
          },
          {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/def-0"
            }
          },
          {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/def-0"
            }
          }
        ],
        "title": "JsonValue"
      }
    }
  },
  "paths": {
...
    "/api/tickets": {
      "get": {
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "post": {
        "operationId": "createTicket",
        "tags": [
          "pilot"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "properties": {
                  "project_id": {
                    "minLength": 1,
                    "type": "string"
                  },
                  "title": {
                    "minLength": 1,
                    "type": "string"
                  },
                  "kind": {
                    "default": "task",
                    "anyOf": [
                      {
                        "anyOf": [
                          {
                            "type": "string",
                            "enum": [
                              "spec"
                            ]
                          },
                          {
                            "type": "string",
                            "enum": [
                              "task"
                            ]
                          },
                          {
                            "type": "string",
                            "enum": [
                              "doc"
                            ]
                          }
                        ]
                      },
                      {
                        "type": "string",
                        "nullable": true,
                        "enum": [
                          null
                        ]
                      }
                    ]
                  },
                  "body": {
                    "default": "",
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "string",
                        "nullable": true,
                        "enum": [
                          null
                        ]
                      }
                    ]
                  },
                  "body_format": {
                    "default": "markdown",
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "markdown"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "html"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          ""
                        ]
                      },
                      {
                        "type": "string",
                        "nullable": true,
                        "enum": [
                          null
                        ]
                      }
                    ]
                  },
                  "priority": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "string",
                        "nullable": true,
                        "enum": [
                          null
                        ]
                      }
                    ]
                  },
                  "labels": {
                    "default": [],
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "parent_id": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "string",
                        "nullable": true,
                        "enum": [
                          null
                        ]
                      }
                    ]
                  },
                  "assignee": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "string",
                        "nullable": true,
                        "enum": [
                          null
                        ]
                      }
                    ]
                  },
                  "created_by": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "string",
                        "nullable": true,
                        "enum": [
                          null
                        ]
                      }
                    ]
                  },
                  "source_ref": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "string",
                        "nullable": true,
                        "enum": [
                          null
                        ]
                      }
                    ]
                  }
                },
                "required": [
                  "project_id",
                  "title"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "seq": {
                      "type": "integer"
                    },
                    "project_id": {
                      "type": "string"
                    },
                    "kind": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "spec"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "task"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "doc"
                          ]
                        }
                      ]
                    },
                    "title": {
                      "type": "string"
                    },
                    "body": {
                      "type": "string"
                    },
                    "body_format": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "markdown"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "html"
                          ]
                        }
                      ]
                    },
                    "body_revision": {
                      "type": "integer"
                    },
                    "state": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "todo"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "in_progress"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "blocked"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "review"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "done"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "archived"
                          ]
                        }
                      ]
                    },
                    "priority": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "string",
                          "nullable": true,
                          "enum": [
                            null
                          ]
                        }
                      ]
                    },
                    "labels": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "parent_id": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "string",
                          "nullable": true,
                          "enum": [
                            null
                          ]
                        }
                      ]
                    },
                    "assignee": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "string",
                          "nullable": true,
                          "enum": [
                            null
                          ]
                        }
                      ]
                    },
                    "created_by": {
                      "type": "string"
                    },
                    "dispatched_to": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "string",
                          "nullable": true,
                          "enum": [
                            null
                          ]
                        }
                      ]
                    },
                    "dispatched_at": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "string",
                          "nullable": true,
                          "enum": [
                            null
                          ]
                        }
                      ]
                    },
                    "source_ref": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "string",
                          "nullable": true,
                          "enum": [
                            null
                          ]
                        }
                      ]
                    },
                    "created_at": {
                      "type": "string"
                    },
                    "updated_at": {
                      "type": "string"
                    },
                    "pseq": {
                      "type": "integer"
                    },
                    "display_id": {
                      "type": "string"
                    },
                    "outline": {
                      "type": "array",
                      "items": {
                        "additionalProperties": true,
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "parent_id": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "string",
                                "nullable": true,
                                "enum": [
                                  null
                                ]
                              }
                            ]
                          },
                          "kind": {
                            "type": "string"
                          },
                          "tag": {
                            "type": "string"
                          },
                          "heading": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "string",
                                "nullable": true,
                                "enum": [
                                  null
                                ]
                              }
                            ]
                          },
                          "short_text": {
                            "type": "string"
                          },
                          "hash": {
                            "type": "string"
                          },
                          "mermaid_error": {
                            "$ref": "#/components/schemas/def-0"
                          }
                        },
                        "required": [
                          "id",
                          "parent_id",
                          "kind",
                          "tag",
                          "heading",
                          "short_text",
                          "hash"
                        ]
                      }
                    },
                    "mermaid_errors": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/def-0"
                      }
                    }
                  },
                  "required": [
                    "id",
                    "seq",
                    "project_id",
                    "kind",
                    "title",
                    "body",
                    "body_format",
                    "body_revision",
                    "state",
                    "priority",
                    "labels",
                    "parent_id",
                    "assignee",
                    "created_by",
                    "dispatched_to",
                    "dispatched_at",
                    "source_ref",
                    "created_at",
                    "updated_at",
                    "pseq",
                    "display_id"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    },
                    "statusCode": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "block_id": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "413": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    },
                    "statusCode": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "block_id": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "415": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    },
                    "statusCode": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "block_id": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    },
                    "statusCode": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "block_id": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  }
}